// src/sdk/model/page-of-eth-address-scan-pipeline.ts
/**
 * UnBlock RESTful Web API
 * An blockchain analysis service
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EthAddressScanPipeline } from './eth-address-scan-pipeline';
import { Pageable } from './pageable';
import { Sort } from './sort';


export interface PageOfEthAddressScanPipeline { 
    content?: Array<EthAddressScanPipeline>;
    first?: boolean;
    last?: boolean;
    number?: number;
    numberOfElements?: number;
    pageable?: Pageable;
    size?: number;
    sort?: Sort;
    totalElements?: number;
    totalPages?: number;
}